backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.1.2
python-multipart
psycopg2-binary
python-dotenv
gunicorn
PyPDF2
boto3
cloudinary
aiofiles
